update for .net 8 and restructure
