+++ .env.example
@@ -1,0 +1,1 @@
+INFINITE_FLIGHT_LIVE_KEY=
