+++ IF-Tools/Components/Layout/MainLayout.razor
@@ -1,5 +1,4 @@
 ﻿@inherits LayoutComponentBase
-@using Microsoft.AspNetCore.Components.Web.Extensions.Head;
 @inject CurrentPage CurrentPage
 
 <Title value="@(CurrentPage.CurrentPageName == "" ? "IF-Tools" : CurrentPage.CurrentPageName + " - IF-Tools")"></Title>
