+++ IF-Tools/Components/Pages/Tools/ActiveATCs.razor
@@ -1,6 +1,4 @@
 @page "/tools/active-atcs"
-@using IFTools.Data.Types
-@using IFTools.Data
 
 @inject CurrentPage CurrentPage
