+++ IF-Tools/Components/Shared/FlightInfoModal.razor
@@ -1,6 +1,3 @@
-@using IFTools.Data.Types
-@using IFTools.Data
-
 @if (_isShown)
 {
     <div class="wrapper">
