+++ IF-Tools/Components/Shared/FrequencyInfoModal.razor
