+++ IF-Tools/Components/_Imports.razor
@@ -1,4 +1,5 @@
 ﻿@using System.Net.Http
+@using static Microsoft.AspNetCore.Components.Web.RenderMode
 @using Microsoft.AspNetCore.Authorization
 @using Microsoft.AspNetCore.Components.Authorization
 @using Microsoft.AspNetCore.Components.Forms
@@ -6,6 +7,8 @@
 @using Microsoft.AspNetCore.Components.Web
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.JSInterop
+
 @using IFTools
-@using IFTools.Shared
-@using IFTools.Shared.Components
+@using IFTools.Components.Shared
+@using IFTools.Data.Types
+@using IFTools.Data
