+++ IF-Tools/IF-Tools.csproj
@@ -1,16 +1,47 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
     <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
+        <TargetFramework>net8.0</TargetFramework>
         <Nullable>disable</Nullable>
         <ImplicitUsings>enable</ImplicitUsings>
         <RootNamespace>IFTools</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="CsvHelper" Version="27.2.1" />
-      <PackageReference Include="Microsoft.AspNetCore.Components.Web.Extensions" Version="5.0.0-preview9.20513.1" />
-      <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+      <PackageReference Include="CsvHelper" Version="32.0.1" />
+      <PackageReference Include="dotenv.net" Version="3.1.3" />
+      <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <_ContentIncludedByDefault Remove="Pages\Error.cshtml" />
+      <_ContentIncludedByDefault Remove="Pages\About.razor" />
+      <_ContentIncludedByDefault Remove="Pages\Index.razor" />
+      <_ContentIncludedByDefault Remove="Pages\Tools\ActiveATCs.razor" />
+      <_ContentIncludedByDefault Remove="Pages\Tools\FlightInfo.razor" />
+      <_ContentIncludedByDefault Remove="Pages\Tools\FlightLookup.razor" />
+      <_ContentIncludedByDefault Remove="Pages\Tools\LiveFlights.razor" />
+      <_ContentIncludedByDefault Remove="Pages\Tools\ServerStatus.razor" />
+      <_ContentIncludedByDefault Remove="Pages\Tools\UserInfo.razor" />
+      <_ContentIncludedByDefault Remove="Pages\Tools\_Root.razor" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <AdditionalFiles Include="Components\Pages\About.razor" />
+      <AdditionalFiles Include="Components\Pages\Error.cshtml" />
+      <AdditionalFiles Include="Components\Pages\Index.razor" />
+      <AdditionalFiles Include="Components\Pages\Tools\ActiveATCs.razor" />
+      <AdditionalFiles Include="Components\Pages\Tools\FlightInfo.razor" />
+      <AdditionalFiles Include="Components\Pages\Tools\FlightLookup.razor" />
+      <AdditionalFiles Include="Components\Pages\Tools\LiveFlights.razor" />
+      <AdditionalFiles Include="Components\Pages\Tools\ServerStatus.razor" />
+      <AdditionalFiles Include="Components\Pages\Tools\UserInfo.razor" />
+      <AdditionalFiles Include="Components\Pages\Tools\_Root.razor" />
+      <AdditionalFiles Include="Components\Shared\FlightInfoModal.razor" />
+      <AdditionalFiles Include="Components\Shared\FrequencyInfoModal.razor" />
+      <AdditionalFiles Include="Components\Shared\InfoDialog.razor" />
+      <AdditionalFiles Include="Components\Shared\LoadingIndicator.razor" />
+      <AdditionalFiles Include="Components\Shared\NotFoundPage.razor" />
     </ItemGroup>
 
 </Project>
