--- IF-Tools/Shared/Components/NavMenu.razor
@@ -1,51 +1,0 @@
-﻿<div class="top-row navbar navbar-dark">
-    <a class="navbar-brand" href="">
-        <img class="logo-image" alt="logo" src="images/logos/logo-white.png">
-    </a>
-    <button class="navbar-toggler" onclick="toggleNavMenu()">
-        <span class="navbar-toggler-icon"></span>
-    </button>
-</div>
-
-<div class="nav-menu collapse" onclick="toggleNavMenu()">
-    <ul class="nav flex-column">
-        <li class="nav-item px-3">
-            <NavLink class="nav-link" href="" Match="NavLinkMatch.All">
-                <span class="feather icon-home" aria-hidden="true"></span> Main
-            </NavLink>
-        </li>
-        <li class="nav-item px-3">
-            <ul class="nav nav-group">
-                <li class="nav-item px-3">
-                    <NavLink class="nav-link" href="/tools/flight-lookup" Match="NavLinkMatch.All">
-                        <span class="feather icon-search" aria-hidden="true"></span> Flight Lookup
-                    </NavLink>
-                </li>
-                <li class="nav-item px-3">
-                    <NavLink class="nav-link" href="/tools/live-flights" Match="NavLinkMatch.All">
-                        <span class="feather icon-navigation" aria-hidden="true"></span> Live Flights
-                    </NavLink>
-                </li>
-                <li class="nav-item px-3">
-                    <NavLink class="nav-link" href="/tools/active-atcs" Match="NavLinkMatch.All">
-                        <span class="feather icon-headphones" aria-hidden="true"></span> Active ATCs
-                    </NavLink>
-                </li>
-                <li class="nav-item px-3">
-                    <NavLink class="nav-link" href="/tools/server-status" Match="NavLinkMatch.All">
-                        <span class="feather icon-server" aria-hidden="true"></span> Server Status
-                    </NavLink>
-                </li>
-                <li class="nav-item px-3">
-                    <NavLink class="nav-link" href="/tools/user-info" Match="NavLinkMatch.Prefix">
-                        <span class="feather icon-user" aria-hidden="true"></span> User Info
-                    </NavLink>
-                </li>
-            </ul>
-        </li>
-    </ul>
-    <div class="nav-bottom-bar">
-        <a class="about-icon" href="/about"><span class="feather icon-info"></span></a>
-        <a href="https://github.com/if-tools/IF-Tools" target="_blank"><span class="feather icon-github"></span></a>
-    </div>
-</div>
