+++ global.json
@@ -1,7 +1,5 @@
 {
   "sdk": {
-    "version": "6.0.200",
-    "rollForward": "latestMajor",
-    "allowPrerelease": true
+    "version": "8.0.204"
   }
 }
